[#2846] Extracted the agent surface into 'AgentSurfaceTrait'.
The questions are declared by the build rather than by a verb, so '--schema', '--validate' and '--agent-help' answer identically wherever the trait is mounted. A bare invocation resolves to a different verb depending on the target directory, and an agent must be able to describe the questions either way.
